# openapi.json
{"openapi":"3.1.0","info":{"title":"APC Clinic API","description":"API for APC Clinic Management System","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/db":{"get":{"summary":"Db Health Check","operationId":"db_health_check_health_db_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/register/":{"post":{"summary":"Register User","operationId":"register_user_user_register__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/login/":{"post":{"summary":"Login User","operationId":"login_user_user_login__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/details/register/":{"post":{"summary":"Register Patient","operationId":"register_patient_patient_details_register__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/details/update/{patient_id}/":{"put":{"summary":"Update Patient Path","description":"Update patient by path parameter ID. Only provided fields will be updated.","operationId":"update_patient_path_patient_details_update__patient_id___put","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/details/update/":{"post":{"summary":"Update Patient Body","description":"Update patient by sending an object containing `id` and the fields to update.","operationId":"update_patient_body_patient_details_update__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/details/{patient_id}/":{"patch":{"summary":"Update Patient Patch","description":"Handle PATCH requests to update partial patient fields at the path the frontend uses.","operationId":"update_patient_patch_patient_details__patient_id___patch","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/treatement/details/register/":{"post":{"tags":["Patient"],"summary":"Register Treatment","operationId":"register_treatment_patient_treatement_details_register__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TreatmentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/treatement/history/{patient_id}":{"get":{"tags":["Patient"],"summary":"Get Treatment History","operationId":"get_treatment_history_patient_treatement_history__patient_id__get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patient/list/":{"get":{"tags":["Patient"],"summary":"Get Patient List","operationId":"get_patient_list_patient_list__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/masters/treatment_charges/":{"get":{"summary":"Get Treatment Charges","operationId":"get_treatment_charges_masters_treatment_charges__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"object"},"type":"array","title":"Response Get Treatment Charges Masters Treatment Charges  Get"}}}}}},"post":{"summary":"Create Treatment Charge","operationId":"create_treatment_charge_masters_treatment_charges__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TreatmentChargeCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Treatment Charge Masters Treatment Charges  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PatientRegister":{"properties":{"name":{"type":"string","title":"Name"},"phone_number":{"type":"string","title":"Phone Number"},"age":{"type":"integer","title":"Age","default":0},"address":{"type":"string","title":"Address","default":""},"city":{"type":"string","title":"City","default":""},"pincode":{"type":"string","title":"Pincode","default":""},"mode_of_referral":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mode Of Referral"}},"type":"object","required":["name","phone_number"],"title":"PatientRegister"},"PatientUpdate":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"pincode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pincode"},"mode_of_referral":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mode Of Referral"}},"type":"object","title":"PatientUpdate"},"TreatmentChargeCreate":{"properties":{"treatment_name":{"type":"string","title":"Treatment Name"},"cost":{"type":"number","title":"Cost"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["treatment_name","cost"],"title":"TreatmentChargeCreate"},"TreatmentCreate":{"properties":{"patient_id":{"type":"string","title":"Patient Id"},"diagnosis":{"type":"string","title":"Diagnosis"},"treatment_plan":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Treatment Plan"},"doctor_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doctor Name"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"items":{"items":{"$ref":"#/components/schemas/TreatmentItemCreate"},"type":"array","title":"Items","default":[]}},"type":"object","required":["patient_id","diagnosis"],"title":"TreatmentCreate"},"TreatmentItemCreate":{"properties":{"treatment_name":{"type":"string","title":"Treatment Name"},"cost":{"type":"number","title":"Cost"}},"type":"object","required":["treatment_name","cost"],"title":"TreatmentItemCreate"},"UserRegister":{"properties":{"username":{"type":"string","title":"Username"},"password_hash":{"type":"string","title":"Password Hash"}},"type":"object","required":["username","password_hash"],"title":"UserRegister"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}